Fix <!DOCTYPE doctype html> bug
According to the HTML specification (see link below), the `DOCTYPE` part of
the DTD is case-insensitive. So for example these DTDs are valid:

    <!doctype html>
    <!DocType html>

But there is a bug in either BeautifulSoup or html.parser: If a document
contains e.g. `<!doctype html>`, then the invalid DTD `<!DOCTYPE doctype html>`
is emitted. This commit adds a somewhat hacky fix for those cases.

https://www.w3.org/TR/html51/syntax.html#the-doctype