Implement injection by link
Instead of injecting tens or hundreds of kilobytes of JavaScript into
every single response, we can now just insert a <script> tag referencing
the userscript, as long as it has a @downloadURL specified.

The @downloadURL is used as the src attribute of the <script> tag. If
the userscript has a @version, it is appended at the end to enable
updates using cache busting. (Note that, as of this commit, there is no
automatic update mechanism; the version suffix is based solely on the
@version in the .user.js file parsed by Userscript Proxy.)