Improve test userscript
I've deliberately not given it a @downloadURL, because that might cause
severe confusion if that URL isn't accessible. The test should only
require Userscript Proxy and example.com in order to work. This also
means that the test userscript has to be injected inline, so I want it
to be small.