Add feature to bypass Content Security Policy (#7)
As pointed out by @deatondg in #6, some sites have a CSP that prevents
userscripts from running properly. This PR makes it possible to either
bypass the CSP specifically for any userscripts that are injected
(inline or not) or disable the CSP altogether whenever a userscript is
injected. The latter is often necessary because userscripts tend to
inject at least one resource into the page, be it an external image or
just some inline CSS.

Resolves #6.

Co-authored-by: deatondg <deaton.dg@gmail.com>