Move typechecking to `builder` stage
As noted by @jurekl, we unnecessarily ship the `typecheck` script and mypy's cache with the packaged application.

This commit reorganizes the Dockerfile so that typechecking is performed in the `builder` stage instead, thereby decreasing the size of the final image by some 10 MB.

However, some unnecessary bloat like the `mypy` library itself is still included in the final image because it's copied over by this line:

    COPY --from=builder /root/.local/lib /root/.local/lib

So further improvements are likely possible.

`ENV PATH=/root/.local/bin:$PATH` is needed in the `builder` stage because otherwise `make start` fails like this at build time:

    ./typecheck: line 7: mypy: command not found

It's also still needed in the final stage because otherwise `make start` fails like this at run time:

    [Errno 2] No such file or directory: 'mitmdump'

💡 `git show --color-moved=plain`